+++ lib/genesis_ruby/api/constants/i18n.rb
@@ -1,0 +1,90 @@
+module GenesisRuby
+  module Api
+    module Constants
+      # WPF localization constants
+      class I18n
+
+        extend Mixins::Constants::Common
+
+        # Arabic
+        AR = 'ar'.freeze
+
+        # Bulgarian
+        BG = 'bg'.freeze
+
+        # German
+        DE = 'de'.freeze
+
+        # English
+        EN = 'en'.freeze
+
+        # Spanish
+        ES = 'es'.freeze
+
+        # French
+        FR = 'fr'.freeze
+
+        # Hindu
+        HI = 'hi'.freeze
+
+        # Japanese
+        JA = 'ja'.freeze
+
+        # Icelandic
+        IS = 'is'.freeze
+
+        # Italian
+        IT = 'it'.freeze
+
+        # Dutch
+        NL = 'nl'.freeze
+
+        # Portuguese
+        PT = 'pt'.freeze
+
+        # Polish
+        PL = 'pl'.freeze
+
+        # Russian
+        RU = 'ru'.freeze
+
+        # Turkish
+        TR = 'tr'.freeze
+
+        # Mandarin Chinese
+        ZH = 'zh'.freeze
+
+        # Indonesian
+        ID = 'id'.freeze
+
+        # Malay
+        MS = 'ms'.freeze
+
+        # Thai
+        TH = 'th'.freeze
+
+        # Czech
+        CS = 'cs'.freeze
+
+        # Croatian
+        HR = 'hr'.freeze
+
+        # Slovenian
+        SL = 'sl'.freeze
+
+        # Finnish
+        FI = 'fi'.freeze
+
+        # Norwegian
+        NO = 'no'.freeze
+
+        # Danish
+        DA = 'da'.freeze
+
+        # Swedish
+        SV = 'sv'.freeze
+
+      end
+    end
+  end
+end