+++ lib/genesis_ruby/api/requests/wpf/create.rb
@@ -1,5 +1,7 @@
 require 'genesis_ruby/utils/transactions/wpf_types'
 require 'genesis_ruby/api/constants/transactions/parameters/sca_exemptions'
+require 'genesis_ruby/api/constants/i18n'
+require 'genesis_ruby/api/mixins/requests/wpf_reminders_attributes'
 
 module GenesisRuby
   module Api
@@ -19,10 +21,17 @@ module GenesisRuby
           include Mixins::Requests::Financial::DynamicDescriptorAttributes
           include Mixins::Requests::Financial::RiskAttributes
           include Mixins::Requests::Financial::Business::BusinessAttributes
+          include Mixins::Requests::WpfRemindersAttributes
 
           attr_reader   :locale, :sca_preference, :sca_exemption
           attr_accessor :transaction_id, :usage, :description, :consumer_id, :return_cancel_url, :remember_card,
-                        :lifetime, :web_payment_form_id
+                        :web_payment_form_id
+
+          MAX_LIFETIME     = 131_487
+          DEFAULT_LIFETIME = 30
+
+          private_constant :MAX_LIFETIME
+          private_constant :DEFAULT_LIFETIME
 
           # The transaction types that the merchant is willing to accept payments for
           def add_transaction_type(name, custom_attributes = {})
@@ -37,7 +46,12 @@ module GenesisRuby
 
           # Define ISO 639-1 language code to the URL
           def locale=(value)
-            init_api_wpf_configuration(language: value.to_s.downcase)
+            allowed_options attribute: __method__,
+                            allowed: Constants::I18n.all,
+                            value: value.to_s.downcase,
+                            allow_empty: true
+
+            init_api_wpf_configuration(language: locale)
           end
 
           # Signifies whether to perform SCA on the transaction. At least one 3DS transaction type has to be submitted.
@@ -57,6 +71,41 @@ module GenesisRuby
                             allowed:     [sca_exemptions::LOW_VALUE, sca_exemptions::LOW_RISK],
                             value:       value.to_s.empty? ? nil : value.to_s.downcase,
                             allow_empty: true
+          end
+
+          # Signifies whether the ’Pay Later’ feature would be enabled on the WPF
+          def pay_later
+            @pay_later ||= false
+
+            return nil unless @pay_later
+
+            @pay_later
+          end
+
+          # Signifies whether the ’Pay Later’ feature would be enabled on the WPF
+          def pay_later=(value)
+            allowed_options attribute: __method__,
+                            allowed: [true, false],
+                            value: value,
+                            allow_empty: true,
+                            error_message: 'Accepts only boolean values'
+          end
+
+          # A number of minutes determining how long the WPF will be valid. Will be set to 30 minutes by default.
+          def lifetime
+            @lifetime ||= DEFAULT_LIFETIME
+          end
+
+          # A number of minutes determining how long the WPF will be valid. Will be set to 30 minutes by default.
+          def lifetime=(value)
+            lifetime = value.to_i
+
+            if lifetime < 1 || lifetime > MAX_LIFETIME
+              raise InvalidArgumentError,
+                    "Valid value ranges between 1 minute and 3 months (#{MAX_LIFETIME} minutes) given in minutes"
+            end
+
+            @lifetime = lifetime
           end
 
           protected
@@ -85,6 +134,8 @@ module GenesisRuby
             unless GenesisRuby::Api::Constants::Currencies::Iso4217.valid?(currency)
               raise ParameterError, "Invalid Currency given with value #{currency}"
             end
+
+            validate_reminders
 
             super
           end
@@ -121,7 +172,9 @@ module GenesisRuby
                 recurring_category:        recurring_category,
                 dynamic_descriptor_params: dynamic_descriptor_structure,
                 risk_params:               risk_parameters_structure,
-                account_owner:             account_owner_attributes_structure
+                account_owner:             account_owner_attributes_structure,
+                pay_later:                 pay_later,
+                reminders:                 pay_later ? reminders_structure : []
               }
             }
           end
@@ -135,6 +188,20 @@ module GenesisRuby
             @transaction_types ||= []
 
             @transaction_types.push(value)
+          end
+
+          # Validate Reminders against the lifetime
+          def validate_reminders
+            return if lifetime.nil?
+
+            reminders.each do |reminder|
+              next unless reminder.after >= lifetime
+
+              raise(
+                ParameterError,
+                "Reminder (#{reminder.after} min) could not be greater than or equal to lifetime (#{lifetime} min)."
+              )
+            end
           end
 
         end