+++ spec/genesis_ruby/api/constants/i18n_spec.rb
@@ -1,0 +1,8 @@
+require 'genesis_ruby/api/constants/i18n'
+require 'spec/genesis_ruby/shared_examples/api/constants/common_examples'
+
+RSpec.describe GenesisRuby::Api::Constants::I18n do
+
+  include_examples 'common constants examples', 'en'
+
+end