+++ spec/genesis_ruby/api/constants/transactions/parameters/wpf/reminder_channels_spec.rb
@@ -1,0 +1,8 @@
+require 'genesis_ruby/api/constants/transactions/parameters/wpf/reminder_channels'
+require 'spec/genesis_ruby/shared_examples/api/constants/common_examples'
+
+RSpec.describe GenesisRuby::Api::Constants::Transactions::Parameters::Wpf::ReminderChannels do
+
+  include_examples 'common constants examples', 'sms'
+
+end