+++ spec/genesis_ruby/stubs/api/requests/wpf_reminders_attributes_stub.rb
@@ -1,0 +1,15 @@
+module GenesisSpec
+  module Stubs
+    module Api
+      module Requests
+        # WPF Reminders Attributes Stub
+        class WpfRemindersAttributesStub
+
+          include GenesisRuby::Api::Mixins::Requests::WpfRemindersAttributes
+          include GenesisRuby::Api::Mixins::Requests::RestrictedSetter
+
+        end
+      end
+    end
+  end
+end